+++ src/main/java/lotto/model/LottoNumbers.java
@@ -7,6 +7,9 @@ import java.util.SortedSet;
 import java.util.TreeSet;
 import java.util.stream.Collectors;
 
+import lotto.model.dto.Number;
+import lotto.model.dto.enums.MatchCount;
+
 public class LottoNumbers {
     static final int NUMBER_SIZE = 6;
     static final String NUMBER_SIZE_ERR_MSG = "로또 숫자는 " + NUMBER_SIZE + "개로 구성되어야 합니다.";
