+++ src/main/java/lotto/model/NumberSupplier.java
@@ -2,6 +2,8 @@ package lotto.model;
 
 import java.util.Set;
 
+import lotto.model.dto.Number;
+
 @FunctionalInterface
 public interface NumberSupplier {
     Set<Number> getNumbers();
