+++ src/main/java/lotto/model/dto/Payment.java
@@ -1,9 +1,9 @@
-package lotto.model;
+package lotto.model.dto;
 
 import java.util.Objects;
 
 public class Payment {
-    static final int LOTTO_PRICE = 1000;
+    public static final int LOTTO_PRICE = 1000;
     private static final String MIN_MONEY_ERR_MSG = "구입 금액은 " + LOTTO_PRICE + "원 이상이어야 합니다.";
     private static final String MONEY_MULTIPLE_ERR_MSG = "구입 금액은 " + LOTTO_PRICE + "의 배수여야 합니다.";
