+++ src/main/java/lotto/model/dto/enums/MatchCount.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.dto.enums;
 
 import java.util.Arrays;
 import java.util.Map;
