+++ src/test/java/lotto/model/LottoNumberGeneratorTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import lotto.model.dto.Payment;
+
 public class LottoNumberGeneratorTest {
     private static final int PAYMENT = 14000;
