+++ src/test/java/lotto/model/LottoNumbersTest.java
@@ -10,6 +10,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import lotto.model.dto.enums.MatchCount;
+
 public class LottoNumbersTest {
     @ParameterizedTest
     @DisplayName(NUMBER_SIZE + "개의 숫자로 구성되어 있지 않을 경우 예외 발생")
