+++ src/test/java/lotto/model/RandomNumberSupplierTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import lotto.model.RandomNumberSupplier.RandomNumberSupplierBuilder;
+import lotto.model.dto.Number;
 
 public class RandomNumberSupplierTest {
     @ParameterizedTest
