+++ src/test/java/lotto/model/dto/enums/MatchCountTest.java
@@ -1,8 +1,10 @@
-package lotto.model;
+package lotto.model.dto.enums;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
+
+import lotto.model.dto.enums.MatchCount;
 
 public class MatchCountTest {
     @Test
